Move helper methods to utils.js
* Move getCaps and insertAppiumPrefixes into utils
* Rename 'getCaps' to 'parseCapsForInnerDriver' so that the purpose of the function is more clear
* Moved 'BASE_CAPS' etc... into helpers.js
