+++ test/helpers.js
@@ -1,6 +1,7 @@
 import path from 'path';
 import wd from 'wd';
 import B from 'bluebird';
+import {insertAppiumPrefixes} from '../lib/utils';
 
 const TEST_HOST = 'localhost';
 const TEST_PORT = 4723;
@@ -24,4 +25,11 @@ function initSession (caps) {
   });
 }
 
-export { initSession, TEST_FAKE_APP, TEST_HOST, TEST_PORT };
+const BASE_CAPS = {platformName: 'Fake', deviceName: 'Fake', app: TEST_FAKE_APP};
+const W3C_PREFIXED_CAPS = {...insertAppiumPrefixes(BASE_CAPS)};
+const W3C_CAPS = {
+  alwaysMatch:{...W3C_PREFIXED_CAPS},
+  firstMatch: [{}],
+};
+
+export { initSession, TEST_FAKE_APP, TEST_HOST, TEST_PORT, BASE_CAPS, W3C_PREFIXED_CAPS, W3C_CAPS };
